+++ spec/support/schema.rb
@@ -3,6 +3,7 @@ ActiveRecord::Schema.define(version: 1) do
     t.string :make
     t.string :model
     t.integer :state, default: 0
+    t.boolean :inspected, default: false
     t.timestamps
   end
 end